Conponentize tipTap and UI fixes
