+++ Clients/src/presentation/components/Table/index.css
@@ -131,3 +131,4 @@
   padding-left: 12px;
   padding-right: 12px;
 }
+
